Working on Upgrade menu
+UpgraderPanel become abstract
+1 Panel script per upgrade type
+Interfce for each type of upgrade

+++ Assets/Scripts/UpgradeMenu/Interface/IUpgradebleSpell.cs
@@ -3,5 +3,5 @@ public interface IUpgradebleSpell
     public void OnRadiusUpgrade(float _value);
     public void OnDamageUpgrade(float _value);
     public void OnAttackRateUpgrade(float _value);
-    public void OnTargetCountUpgrade(int _increment);
+    public void OnTargetCountUpgrade(float _increment);
 }
